Update to TouchGFX 4.20

+++ CM7/Middlewares/ST/touchgfx/framework/include/touchgfx/TextureMapTypes.hpp
@@ -2,7 +2,7 @@
 * Copyright (c) 2018(-2022) STMicroelectronics.
 * All rights reserved.
 *
-* This file is part of the TouchGFX 4.19.1 distribution.
+* This file is part of the TouchGFX 4.20.0 distribution.
 *
 * This software is licensed under terms that can be found in the LICENSE file in
 * the root directory of this software component.
@@ -138,7 +138,7 @@ struct Edge
  */
 FORCE_INLINE_FUNCTION float fixed28_4ToFloat(fixed28_4 value)
 {
-    return value / 16.0f;
+    return (float)value / 16.0f;
 }
 
 /**
